ci: implement auto release chain
