+++ .github/workflows/deploy.yml
@@ -2,6 +2,8 @@ name: Deploy
 on:
   workflow_dispatch:
 
+  workflow_call:
+    
 jobs:
   deploy-it:
     runs-on: ubuntu-latest
@@ -10,7 +12,7 @@ jobs:
         with:
           fetch-depth: 0
           token: ${{ secrets.PAT }}
-
+          
       - name: Prepare
         uses: ./.github/actions/prepare
