--- .github/workflows/release.yml
@@ -1,24 +1,0 @@
-name: Release
-on:
-  workflow_dispatch:
-
-jobs:
-  release:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-          token: ${{ secrets.PAT }}
-
-      - name: Prepare
-        uses: ./.github/actions/prepare
-
-      - name: Release Production
-        run: yarn run release-prod
-        env:
-          REF: ${{ github.ref }}
-          GH_TOKEN: ${{ github.token }}
-          # Github will create .npmrc based on this env variable.
-          # https://docs.github.com/en/actions/publishing-packages/publishing-nodejs-packages#publishing-packages-to-the-npm-registry
-          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
